Add evenlistener to the thumbnail at thumbnails.js. Also renamed file to the show-big-picture and added code for big picture open\close.  Deleted from main.js unusable import .
