+++ js/main.js
@@ -1,5 +1,4 @@
 import { getPictures } from './data.js';
 import { makeAllThumbnail } from './thumbnails.js';
-import { makeBigPicture } from './bigPicture-view.js';
 
 makeAllThumbnail(getPictures());
